split ParseOptions into copybook-parse-options microcrate
