+++ crates/copybook-core/Cargo.toml
@@ -28,6 +28,7 @@ copybook-overflow.workspace = true
 copybook-utils.workspace = true
 copybook-dialect.workspace = true
 copybook-lexer.workspace = true
+copybook-parse-options.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_plain.workspace = true
