+++ crates/copybook-parse-options/Cargo.toml
@@ -1,0 +1,24 @@
+# SPDX-License-Identifier: AGPL-3.0-or-later
+[package]
+name = "copybook-parse-options"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Shared parse option contracts for copybook parsing crates."
+documentation = "https://docs.rs/copybook-parse-options"
+readme = "README.md"
+keywords = ["cobol", "copybook", "parser", "options"]
+categories = ["parsing", "config"]
+
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE"]
+
+[lints]
+workspace = true
+
+[dependencies]
+copybook-dialect.workspace = true
+serde.workspace = true
